test: merge serial tests in cmd, planner, server, util (#31003)
