+++ cmd/ddltest/index_test.go
