+++ server/statistics_handler_test.go
